car_rent/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-3-jet==1.0.8
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
